moved 'joinTeam' form to own component, team is chosen from select element of all teams in system
